wip: extract KeyValuePairStore
