+++ test/phpunit/FormDataTest.php
@@ -1,0 +1,8 @@
+<?php
+namespace Gt\Http\Test;
+
+use PHPUnit\Framework\TestCase;
+
+class FormDataTest extends TestCase {
+
+}
